+++ lib/theme/ui/profile/own_profile.dart
@@ -1,11 +1,13 @@
 import 'dart:async';
 
+import 'package:chat_interface/controller/current/connection_controller.dart';
 import 'package:chat_interface/controller/spaces/spaces_controller.dart';
 import 'package:chat_interface/controller/current/status_controller.dart';
 import 'package:chat_interface/database/database.dart';
 import 'package:chat_interface/main.dart';
 import 'package:chat_interface/pages/chat/sidebar/friends/friends_page.dart';
 import 'package:chat_interface/pages/settings/data/settings_controller.dart';
+import 'package:chat_interface/pages/status/setup/setup_manager.dart';
 import 'package:chat_interface/theme/components/forms/icon_button.dart';
 import 'package:chat_interface/theme/ui/dialogs/window_base.dart';
 import 'package:chat_interface/theme/ui/profile/developer_window.dart';
@@ -275,7 +277,7 @@ class _ProfileState extends State<OwnProfile> {
           ),
           verticalSpacing(elementSpacing),
 
-          //* Hide profile
+          // For debug only database viewer
           if (isDebug)
             Padding(
               padding: const EdgeInsets.only(bottom: elementSpacing),
@@ -286,6 +288,20 @@ class _ProfileState extends State<OwnProfile> {
                   testLoading.value = true;
                   unawaited(Navigator.of(context).push(MaterialPageRoute(builder: (context) => DriftDbViewer(db))));
                   testLoading.value = false;
+                },
+                loading: testLoading,
+              ),
+            ),
+
+          // For debug only retry button
+          if (isDebug)
+            Padding(
+              padding: const EdgeInsets.only(bottom: elementSpacing),
+              child: ProfileButton(
+                icon: Icons.restart_alt,
+                label: 'profile.retry'.tr,
+                onTap: () async {
+                  Get.find<ConnectionController>().restart();
                 },
                 loading: testLoading,
               ),